Refactor into pip installable package
