+++ pabloSim/Grover.py
@@ -6,7 +6,7 @@ register.
 """
 # From ListQuantumGates we import the number of qubits of the quantum register (n)
 # and the needed quantum gates.
-from ListQuantumGates import n, Hadamard, O, G
+from .ListQuantumGates import n, Hadamard, O, G
 import numpy as np
 
 def main():
