+++ stewartSim/ListQuantumGates.py
@@ -3,7 +3,7 @@
 Uses the Quantum_Gate class to create all the quantum gates that the quantum
 computer will need.
 """
-from Quantum_Gate import Quantum_Gate as QGate
+from .Quantum_Gate import Quantum_Gate as QGate
 
 import numpy as np 
 import sys
