--- template/requirements.txt
@@ -1,3 +1,0 @@
-setuptools
-scipy
-numpy
