--- template/template_package/console_entry_point.py
@@ -1,3 +1,0 @@
-def main():
-    print("This script has been run")
-
